# src/main/java/com/example/task/management/system/model/request/NewCommentRequest.java
package com.example.task.management.system.model.request;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Builder
@Data
@AllArgsConstructor
@NoArgsConstructor
@Schema(description = "Запрос на создание комментария")
public class NewCommentRequest {
    @NotBlank
    private String text;
    @NotNull
    private Long taskId;
}