# src/main/java/com/example/task/management/system/model/response/TasksListResponse.java
package com.example.task.management.system.model.response;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@Schema(description = "Ответ со всеми задачами")
public class TasksListResponse {
    private List<TaskResponse> allTasksList;
}